Polish apache/dubbo#6389 : [Issue] Resolving the issues with ConsulServiceDiscovery
