+++ dubbo-common/src/test/java/org/apache/dubbo/common/convert/multiple/StringToListConverterTest.java
@@ -14,16 +14,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.dubbo.convert.multiple;
+package org.apache.dubbo.common.convert.multiple;
 
-import org.apache.dubbo.common.convert.multiple.StringToQueueConverter;
 import org.apache.dubbo.common.utils.CollectionUtils;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.util.AbstractList;
-import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Deque;
@@ -40,23 +38,24 @@ import java.util.concurrent.ConcurrentSkipListSet;
 import java.util.concurrent.TransferQueue;
 
 import static java.util.Arrays.asList;
+import static org.apache.dubbo.common.extension.ExtensionLoader.getExtensionLoader;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
- * {@link StringToQueueConverter} Test
+ * {@link StringToListConverter} Test
  *
  * @since 2.7.6
  */
-public class StringToQueueConverterTest {
+public class StringToListConverterTest {
 
-    private StringToQueueConverter converter;
+    private MultiValueConverter converter;
 
     @BeforeEach
     public void init() {
-        converter = new StringToQueueConverter();
+        converter = getExtensionLoader(MultiValueConverter.class).getExtension("string-to-list");
     }
 
     @Test
@@ -64,22 +63,22 @@ public class StringToQueueConverterTest {
 
         assertFalse(converter.accept(String.class, Collection.class));
 
-        assertFalse(converter.accept(String.class, List.class));
-        assertFalse(converter.accept(String.class, AbstractList.class));
+        assertTrue(converter.accept(String.class, List.class));
+        assertTrue(converter.accept(String.class, AbstractList.class));
         assertTrue(converter.accept(String.class, LinkedList.class));
-        assertFalse(converter.accept(String.class, ArrayList.class));
-
-        assertTrue(converter.accept(String.class, Queue.class));
-        assertTrue(converter.accept(String.class, BlockingQueue.class));
-        assertTrue(converter.accept(String.class, TransferQueue.class));
-        assertTrue(converter.accept(String.class, Deque.class));
-        assertTrue(converter.accept(String.class, BlockingDeque.class));
+        assertTrue(converter.accept(String.class, ArrayList.class));
 
         assertFalse(converter.accept(String.class, Set.class));
         assertFalse(converter.accept(String.class, SortedSet.class));
         assertFalse(converter.accept(String.class, NavigableSet.class));
         assertFalse(converter.accept(String.class, TreeSet.class));
         assertFalse(converter.accept(String.class, ConcurrentSkipListSet.class));
+
+        assertFalse(converter.accept(String.class, Queue.class));
+        assertFalse(converter.accept(String.class, BlockingQueue.class));
+        assertFalse(converter.accept(String.class, TransferQueue.class));
+        assertFalse(converter.accept(String.class, Deque.class));
+        assertFalse(converter.accept(String.class, BlockingDeque.class));
 
         assertFalse(converter.accept(null, char[].class));
         assertFalse(converter.accept(null, String.class));
@@ -90,16 +89,15 @@ public class StringToQueueConverterTest {
     @Test
     public void testConvert() {
 
-        Queue values = new ArrayDeque(asList(1.0, 2.0, 3.0));
+        List values = asList(1, 2, 3);
 
-        Queue result = (Queue<Double>) converter.convert("1.0,2.0,3.0", Queue.class, Double.class);
+        List result = (List<Integer>) converter.convert("1,2,3", List.class, Integer.class);
 
         assertTrue(CollectionUtils.equals(values, result));
 
-        values.clear();
-        values.add(123);
+        values = asList("123");
 
-        result = (Queue) converter.convert("123", Queue.class, Integer.class);
+        result = (List<String>) converter.convert("123", List.class, String.class);
 
         assertTrue(CollectionUtils.equals(values, result));
