+++ dubbo-common/src/test/java/org/apache/dubbo/common/convert/multiple/StringToTransferQueueConverterTest.java
@@ -14,10 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.dubbo.convert.multiple;
+package org.apache.dubbo.common.convert.multiple;
 
-import org.apache.dubbo.common.convert.multiple.MultiValueConverter;
-import org.apache.dubbo.common.convert.multiple.StringToListConverter;
 import org.apache.dubbo.common.utils.CollectionUtils;
 
 import org.junit.jupiter.api.BeforeEach;
@@ -37,6 +35,7 @@ import java.util.TreeSet;
 import java.util.concurrent.BlockingDeque;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.ConcurrentSkipListSet;
+import java.util.concurrent.LinkedTransferQueue;
 import java.util.concurrent.TransferQueue;
 
 import static java.util.Arrays.asList;
@@ -47,17 +46,17 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
- * {@link StringToListConverter} Test
+ * {@link StringToTransferQueueConverter} Test
  *
  * @since 2.7.6
  */
-public class StringToListConverterTest {
+public class StringToTransferQueueConverterTest {
 
     private MultiValueConverter converter;
 
     @BeforeEach
     public void init() {
-        converter = getExtensionLoader(MultiValueConverter.class).getExtension("string-to-list");
+        converter = getExtensionLoader(MultiValueConverter.class).getExtension("string-to-transfer-queue");
     }
 
     @Test
@@ -65,10 +64,10 @@ public class StringToListConverterTest {
 
         assertFalse(converter.accept(String.class, Collection.class));
 
-        assertTrue(converter.accept(String.class, List.class));
-        assertTrue(converter.accept(String.class, AbstractList.class));
-        assertTrue(converter.accept(String.class, LinkedList.class));
-        assertTrue(converter.accept(String.class, ArrayList.class));
+        assertFalse(converter.accept(String.class, List.class));
+        assertFalse(converter.accept(String.class, AbstractList.class));
+        assertFalse(converter.accept(String.class, LinkedList.class));
+        assertFalse(converter.accept(String.class, ArrayList.class));
 
         assertFalse(converter.accept(String.class, Set.class));
         assertFalse(converter.accept(String.class, SortedSet.class));
@@ -78,9 +77,9 @@ public class StringToListConverterTest {
 
         assertFalse(converter.accept(String.class, Queue.class));
         assertFalse(converter.accept(String.class, BlockingQueue.class));
-        assertFalse(converter.accept(String.class, TransferQueue.class));
         assertFalse(converter.accept(String.class, Deque.class));
         assertFalse(converter.accept(String.class, BlockingDeque.class));
+        assertTrue(converter.accept(String.class, TransferQueue.class));
 
         assertFalse(converter.accept(null, char[].class));
         assertFalse(converter.accept(null, String.class));
@@ -91,15 +90,17 @@ public class StringToListConverterTest {
     @Test
     public void testConvert() {
 
-        List values = asList(1, 2, 3);
+        TransferQueue values = new LinkedTransferQueue(asList(1, 2, 3));
 
-        List result = (List<Integer>) converter.convert("1,2,3", List.class, Integer.class);
+        TransferQueue result = (TransferQueue) converter.convert("1,2,3", List.class, Integer.class);
 
         assertTrue(CollectionUtils.equals(values, result));
 
-        values = asList("123");
+        values.clear();
 
-        result = (List<String>) converter.convert("123", List.class, String.class);
+        values.addAll(asList("123"));
+
+        result = (TransferQueue) converter.convert("123", NavigableSet.class, String.class);
 
         assertTrue(CollectionUtils.equals(values, result));
 
@@ -114,6 +115,6 @@ public class StringToListConverterTest {
 
     @Test
     public void testGetPriority() {
-        assertEquals(Integer.MAX_VALUE - 2, converter.getPriority());
+        assertEquals(Integer.MAX_VALUE - 4, converter.getPriority());
     }
 }
